[Add] View to player profile + api refactor
