+++ WynnAPI/guilds/__init__.py
@@ -1,0 +1,1 @@
+from WynnAPI.guilds.get_guild import *
