+++ WynnAPI/players/__init__.py
@@ -1,0 +1,1 @@
+from WynnAPI.players.get_players import *
