+++ WynnAPI/territories/__init__.py
@@ -1,0 +1,1 @@
+from WynnAPI.territories.get_territories import *
