--- guilds/__init__.py
@@ -1,1 +1,0 @@
-from guilds.get_guild import *
