--- players/__init__.py
@@ -1,1 +1,0 @@
-from players.get_players import *
